graphix update

+++ RTOSWorkspace/bitpattern_search/test.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main()
 {
-printf("%d\n",sizeof(unsigned long));
+printf("%d\n",sizeof(/*unsigned long long*/ int));
 return 0;
 }
